Reorganizing stuff based on latest changes in repo
